staging version without ssr

+++ src/app/components/profile/profile-settings/profile-settings.component.ts
@@ -1,22 +1,33 @@
-import { ChangeDetectionStrategy, Component, OnInit, ViewChild, inject } from '@angular/core';
-import { AbstractControlOptions, FormBuilder, ReactiveFormsModule, Validators } from '@angular/forms';
+import {
+  Component,
+  OnInit,
+  PLATFORM_ID,
+  ViewChild,
+  inject,
+} from '@angular/core';
+import {
+  AbstractControlOptions,
+  FormBuilder,
+  ReactiveFormsModule,
+  Validators,
+} from '@angular/forms';
 import { MatchPasswordValidator } from '../../../core/validators/match-password.validator';
 import { PasswordValidator } from '../../../core/validators/password.validator';
-import { FieldComponent } from "../../../shared/components/form/field/field.component";
-import { LabelComponent } from "../../../shared/components/form/label/label.component";
+import { FieldComponent } from '../../../shared/components/form/field/field.component';
+import { LabelComponent } from '../../../shared/components/form/label/label.component';
 import { FormElementDirective } from '../../../shared/directives/form-element.directive';
-import { ErrorComponent } from "../../../shared/components/form/error/error.component";
-import { ButtonComponent } from "../../../shared/components/form/button/button.component";
-import { CommonModule } from '@angular/common';
-import { IconEyeComponent } from "../../../shared/icons/eye/icon-eye.component";
-import { IconEyeOffComponent } from "../../../shared/icons/eye-off/icon-eye-off.component";
+import { ErrorComponent } from '../../../shared/components/form/error/error.component';
+import { ButtonComponent } from '../../../shared/components/form/button/button.component';
+import { CommonModule, isPlatformBrowser } from '@angular/common';
+import { IconEyeComponent } from '../../../shared/icons/eye/icon-eye.component';
+import { IconEyeOffComponent } from '../../../shared/icons/eye-off/icon-eye-off.component';
 import { AuthService } from '../../../core/services/auth/auth.service';
 import { IUser } from '../../../shared/interfaces/auth/user.interface';
-import { IconEditComponent } from "../../../shared/icons/edit/icon-edit.component";
-import { IconSpinnerComponent } from "../../../shared/icons/spinner/icon-spinner.component";
+import { IconEditComponent } from '../../../shared/icons/edit/icon-edit.component';
+import { IconSpinnerComponent } from '../../../shared/icons/spinner/icon-spinner.component';
 import { IApiResponse } from '../../../shared/interfaces/api-response-interface';
 import { MatSnackBar } from '@angular/material/snack-bar';
-import { ModalComponent } from "../../../shared/components/modal/modal.component";
+import { ModalComponent } from '../../../shared/components/modal/modal.component';
 import { FileValidator } from '../../../core/validators/file.validator';
 import { EmailValidator } from '../../../core/validators/email.validator';
 import { StringValidator } from '../../../core/validators/string.validator';
@@ -24,6 +35,7 @@ import { PhoneValidator } from '../../../core/validators/phone.validator';
 import { Meta } from '@angular/platform-browser';
 import { NavigationEnd, Router } from '@angular/router';
 import { filter } from 'rxjs';
+import { MetaService } from '../../../core/services/meta-data/meta.service';
 @Component({
   selector: 'app-profile-settings',
   standalone: true,
@@ -40,10 +52,8 @@ import { filter } from 'rxjs';
     IconEyeComponent,
     IconEyeOffComponent,
     IconEditComponent,
-    IconSpinnerComponent,
     ModalComponent,
   ],
-  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class ProfileSettingsComponent implements OnInit {
   user: IUser;
@@ -51,38 +61,23 @@ export class ProfileSettingsComponent implements OnInit {
   fb: FormBuilder = inject(FormBuilder);
   authService: AuthService = inject(AuthService);
   _snackBar: MatSnackBar = inject(MatSnackBar);
-  metaService: Meta = inject(Meta);
+  _MetaService: MetaService = inject(MetaService);
+  private platformId = inject(PLATFORM_ID);
+  isBrowser = isPlatformBrowser(this.platformId); // Determine if running on browser
   ngOnInit(): void {
-    this.setCanonicalURL(window.location.href);
+    if (this.isBrowser) {
+      this._MetaService.setCanonicalURL(window.location.href);
 
-    // Update the canonical URL on route changes
-    this.router.events
-      .pipe(filter((event) => event instanceof NavigationEnd))
-      .subscribe(() => {
-        this.setCanonicalURL(window.location.href);
-      });
+      this.router.events
+        .pipe(filter((event) => event instanceof NavigationEnd))
+        .subscribe(() => {
+          this._MetaService.setCanonicalURL(window.location.href);
+        });
+    }
 
     this.setAuthUserInfoToEditProfileForm();
   }
 
-  setCanonicalURL(url: string) {
-    let link: HTMLLinkElement =
-      document.querySelector("link[rel='canonical']") || null;
-
-    if (link) {
-      link.setAttribute('href', url);
-    } else {
-      link = document.createElement('link');
-      link.setAttribute('rel', 'canonical');
-      link.setAttribute('href', url);
-      document.head.appendChild(link);
-    }
-    // Set og:url
-    this.metaService.updateTag({
-      property: 'og:url',
-      content: url,
-    });
-  }
   /*
     -----------------------------
     -- Change User Profile Image
@@ -244,7 +239,9 @@ export class ProfileSettingsComponent implements OnInit {
       .subscribe({
         next: (res: IApiResponse<IUser>) => {
           if (res.success) {
-            localStorage.clear();
+            if (this.isBrowser) {
+              localStorage.clear();
+            }
             this.router.navigateByUrl('/login');
             this._snackBar.open('Password changed successfully.', '✖', {
               panelClass: 'success-snackbar',
